pyproject.toml
[tool.poetry]
name = "inventree-organizer-generator"
version = "0.1.0"
description = "Tool to generate locations for organizer boxes in Inventree"
authors = ["Jasper Chan"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.9"
inventree = "^0.6.0"
PyYAML = "^6.0"
fire = "^0.4.0"
pandas = "^1.4.2"
pathvalidate = "^2.5.0"

[tool.poetry.dev-dependencies]
pylint = "^2.13.9"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
